Add operations support
