+++ lib/trailblazer/operation/responder.rb
@@ -1,23 +1,8 @@
-module Trailblazer::Operation::Responder
-  def self.included(base)
-    base.extend ClassMethods
-  end
-
-  module ClassMethods
-    def model_name
-      model_class.model_name
+class Trailblazer::Operation
+  module Responder
+    def errors
+      return [] if valid?
+      [1]
     end
-  end
-
-  extend Forwardable
-  def_delegators :@model, :to_param, :destroyed?, :persisted?
-
-  def errors
-    return [] if @valid
-    [1]
-  end
-
-  def to_model
-    @model
   end
 end
