+++ lib/trailblazer/rails.rb
@@ -9,20 +9,26 @@ end
 require "trailblazer/rails/railtie"
 
 require "trailblazer/operation"
-# TODO: remove that once i18n, validations etc in Reform/AM are sorted.
 Trailblazer::Operation.contract_class.class_eval do
+  # TODO: remove that once i18n, validations etc in Reform/AM are sorted.
   def self.name
     # for whatever reason, validations climb up the inheritance tree and require _every_ class to have a name (4.1).
     "Reform::Form"
   end
-end
 
-# Automatically set model_name on operation's contract when `Op::Model` is included.
-require "trailblazer/operation/model"
-require "trailblazer/operation/model/active_model"
-Trailblazer::Operation::Model::DSL.module_eval do
-  include Trailblazer::Operation::Model::ActiveModel # ::contract.
+  # For modeless operations. It will be override if model is add.
+  def persisted?
+    false
+  end
 end
 
 require "trailblazer/autoloading"
 require "trailblazer/rails/autoloading"
+
+# Automatically set model_name on operation's contract when `Op::Model` is included.
+require "trailblazer/operation/model"
+require "trailblazer/operation/model/active_model"
+require "trailblazer/operation/model/responder"
+Trailblazer::Operation::Model::DSL.module_eval do
+  include Trailblazer::Operation::Model::ActiveModel # ::contract.
+end
