+++ lib/trailblazer/rails/railtie.rb
@@ -2,13 +2,19 @@ require "rails/railtie"
 
 module Trailblazer
   class Railtie < ::Rails::Railtie
-    def self.autoload_operations(app)
+    def self.autoload_cruds(app)
       Dir.glob("app/concepts/**/crud.rb") do |f|
         path  = f.sub("app/concepts/", "")
         model = path.sub("/crud.rb", "")
 
         require_dependency "#{app.root}/app/models/#{model}" # load the model file, first (thing.rb).
         require_dependency "#{app.root}/#{f}" # load app/concepts/{concept}/crud.rb (Thing::Create, Thing::Update, and so on).
+      end
+    end
+
+    def self.autoload_operations(app)
+      Dir.glob("app/concepts/**/operations.rb") do |f|
+        require_dependency "#{app.root}/#{f}" # load app/concepts/{concept}/operations.rb
       end
     end
 
@@ -28,6 +34,7 @@ module Trailblazer
       # the trb autoloading has to be run after initializers have been loaded, so we can tweak inclusion of features in
       # initializers.
       ActionDispatch::Reloader.to_prepare do
+        Trailblazer::Railtie.autoload_cruds(app)
         Trailblazer::Railtie.autoload_operations(app)
         Trailblazer::Railtie.autoload_cells(app)
       end
